迁移到x86_64并修复大量类型错误和内存布局错误；完善对象缓存释放逻辑；printf和vsprintf改用编译器内建va_list；重构页表管理，新增页表锁；修改rust版链表实现，内部使用统一的Link类型；新增读写自旋锁；修复GRUB安装脚本bug

+++ src/arch/x86/bios_emu/includes/stack.h
@@ -1,8 +1,8 @@
 #ifndef _BIOS_EMU_STACK_H
 #define _BIOS_EMU_STACK_H
 
-#define STACK_POINTER16(env) ((env->regs.ss << 4) + env->regs.sp)
-#define STACK_POINTER32(env) env->regs.esp
+#define STACK_POINTER16(env) ((size_t)((env->regs.ss << 4) + env->regs.sp))
+#define STACK_POINTER32(env) ((size_t)env->regs.esp)
 
 #define PUSH16(env, data, size)                                      \
 	{                                                                \
